requirements.txt
gunicorn==19.9.0
psycopg2==2.8.3
pyTelegramBotAPI==3.6.6
requests==2.7.0
six==1.12.0
telebot==0.0.3
certifi==2019.6.16
chardet==3.0.4
get==2019.4.13
idna==2.8
lxml==4.3.4
post==2019.4.13
public==2019.4.13
python-docx==0.8.10
query-string==2019.4.13
request==2019.4.13
urllib3==1.25.3